Implementing a more manageable solution for player sync and fixing inbound packets
