+++ src/net/packets/outbound-packets/index.ts
@@ -6,6 +6,7 @@ import { logoutPacket } from './logout-packet';
 import { welcomeScreenPacket } from './welcome-screen-packet';
 import { friendsListPacket } from './friends-list-packet';
 import { updateSkillPacket } from './update-skill-packet';
+import { playerSyncPacket } from './player-sync/player-sync-packet';
 
 const outboundPackets: OutboundPacket[] = [
     chatboxMessagePacket,
@@ -15,6 +16,7 @@ const outboundPackets: OutboundPacket[] = [
     welcomeScreenPacket,
     friendsListPacket,
     updateSkillPacket,
+    playerSyncPacket,
 ];
 
 export default outboundPackets;
