+++ src/net/packets/outbound-packets/update-map-region-packet.ts
@@ -9,7 +9,8 @@ interface UpdateMapRegionData {
 export const updateMapRegionPacket: OutboundPacket<UpdateMapRegionData> = {
     name: 'updateMapRegion',
     opcodes: {
-        319: 228
+        319: 228,
+        357: 121,
     },
     encoders: {
         319: (player, opcode, data) => {
@@ -17,6 +18,12 @@ export const updateMapRegionPacket: OutboundPacket<UpdateMapRegionData> = {
             buffer.put(data.mapCoords.x, 'short');
             buffer.put(data.mapCoords.y, 'short', 'le');
             return buffer;
-        }
+        },
+        357: (player, opcode, data) => {
+            const buffer = new ByteBuffer(4);
+            buffer.put(data.mapCoords.y, 'short', 'le');
+            buffer.put(data.mapCoords.x, 'short');
+            return buffer;
+        },
     },
 };
